Rewrite from the ground-up.
I spent the last couple of months rewriting Kevlar from the ground up.
There was a lot of experimentation and throw-away code in the Git
history so I decided to squash everything in one giant commit.

The most noticeable change is that now Kevlar uses the Dhall
configuration language for specifying the build steps.  This choice
guided me towards a  much simpler implementation.

[1] https://dhall-lang.org/
[2] http://wiki.c2.com/?PlanToThrowOneAway