Add usingSourceBuiltImage
This function replaces bakeDockerImage and loadFromStep. The function
usingSourceBuiltImage adds a image baker step and modifies the provided
step so it uses the built docker image.

We now pass a whole context directory to the image baker. Interpolating
from the Dhall configuration seemed to be a good idea, but working with
file paths in this case makes more sense.